IoT/requirements.txt
Adafruit-Blinka==8.29.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-requests==2.0.3
adafruit-circuitpython-sht4x==1.0.19
adafruit-circuitpython-typing==1.9.6
adafruit-extended-bus==1.0.2
Adafruit-PlatformDetect==3.57.0
Adafruit-PureIO==1.1.11
ansi2html==1.9.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
Flask==3.0.0
greenponik-sht40==0.0.1
grove.py==0.6
grovepi==1.0.4
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
nest-asyncio==1.5.8
numpy==1.26.3
packaging==23.2
plotly==5.18.0
pycparser==2.21
pyftdi==0.55.0
pyserial==3.5
pyusb==1.2.1
requests==2.31.0
retrying==1.3.4
RPi.GPIO==0.7.1
scipy==1.11.4
seeed-python-dht==0.0.2
six==1.16.0
smbus-cffi==0.5.1
smbus2==0.4.3
spidev==3.6
tenacity==8.2.3
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
